Finish cpop
