+++ src/main/java/at/ac/tuwien/ec/sleipnir/ThesisMain.java
@@ -9,11 +9,10 @@ import at.ac.tuwien.ec.model.software.MobileApplication;
 import at.ac.tuwien.ec.model.software.MobileWorkload;
 import at.ac.tuwien.ec.model.software.mobileapps.WorkloadGenerator;
 import at.ac.tuwien.ec.scheduling.offloading.OffloadScheduling;
-import at.ac.tuwien.ec.scheduling.offloading.algorithms.cpopbased.CPOPRuntime;
+import at.ac.tuwien.ec.scheduling.offloading.algorithms.thesis.cpop.CPOPRuntime;
 import at.ac.tuwien.ec.scheduling.offloading.algorithms.heftbased.HEFTResearch;
 import java.util.ArrayList;
 import scala.Tuple2;
-import scala.Tuple5;
 
 public class ThesisMain {
   public static void main(String[] args) {
